+++ src/components/ClientService/styles.js
@@ -13,32 +13,13 @@ export const Photo = styled.img`
   src: ${(props) => props.src};
 `;
 
-export const Paragraph = styled.p`
-  display: flex;
-  margin: 10px 0 0 0;
-`
-
 export const IntDivider = styled.div`
   display: flex;
   flex: 1;
   flex-direction: column;
   align-items: center;
   justify-content: center;
-`
-export const Information = styled.div`
-  display: flex;
-  flex: 4;
-  width: 300px;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
 `;
-
-export const Meter = styled.meter`
-  width: 100%;
-  height: 20px;
-  margin: 15px 0;
-`
 
 export const ContainerList = styled.article`
   display: flex;
