+++ src/components/ContainerInformationService/styles.js
@@ -1,0 +1,21 @@
+import styled from 'styled-components';
+
+export const Information = styled.div`
+  display: flex;
+  flex: 4;
+  width: 300px;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+`;
+
+export const Meter = styled.meter`
+  width: 100%;
+  height: 20px;
+  margin: 15px 0;
+`;
+
+export const Paragraph = styled.p`
+  display: flex;
+  margin: 10px 0 0 0;
+`;
